reservations-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.38
python-jose[cryptography]==3.5.0
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.28.1
